opus_platform_watchdog/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
httpx>=0.28.0
aiosqlite>=0.20.0
jinja2>=3.1.0
python-dotenv>=1.0.0
